% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{deleteStataFuture}
\alias{deleteStataFuture}
\title{Remove Stata task/job sent to Stata by \code{\link[RStataLink]{doInStata}}}
\usage{
deleteStataFuture(StataFuture)
}
\arguments{
\item{StataFuture}{An object of S3 class 'StataFuture' produced by
\code{\link[RStataLink]{doInStata}}, with argument \code{future = TRUE},
when a job was sent to Stata.}
}
\description{
This is done if possible (if not yet executed or started).
}

