% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{doInStata}
\alias{doInStata}
\title{Execute some Stata code in a running Stata instance}
\usage{
doInStata(id, code = "", df = NULL, import_df = !is.null(df),
  results = c("e", "r"), timeout = Inf, preserve_restore = FALSE,
  cleanup = TRUE, nolog = FALSE, future = FALSE)
}
\arguments{
\item{id}{An object of S3 class 'StataID' generated by \code{\link[RStataLink]{startStata}}}

\item{code}{A Stata code in a string to be executed}

\item{df}{A data frame to be loaded into Stata before the Stata code is executed (optional)}

\item{import_df}{Logical: should the resulting Stata dataset be returned?
Defaults: \code{TRUE} if \code{df} is provided, else \code{FALSE}.}

\item{results}{NULL or a character vector with either \code{"e"} or \code{"r"} or both of these
characters. Default: \code{c("e", "r")} which means that both e-class and r-class Stata results
will be imported (stored in e() and r() Stata macros, scalars, matrices respectively). If
\code{NULL}, none of these results will be imported.}

\item{timeout}{Seconds for R to wait for Stata results. Default: infinity.}

\item{preserve_restore}{Logical: should the Stata code start with \code{preserve} and end with \code{restore}?
Default: \code{FALSE}.}

\item{cleanup}{Logical: should the time-stamped temporary files (.do file, .log file, and possible input
and output .tsv files) be deleted at the end? Default: \code{TRUE}.}

\item{nolog}{Logical: should the Stata log be discarded in the returned value? Default: \code{FALSE}.}

\item{future}{Logical: should this function block R until Stata finishes the job and return what Stata
produces (\code{future = FALSE}, default) or should this function send the job to Stata and
not make R wait for Stata output (\code{future = TRUE}). In latter case, the Stata output can
be obtained with the function \code{\link[RStataLink]{getStataFuture}}.}
}
\value{
If \code{future = FALSE}, a list with:
\itemize{
         \item \code{log} -- A character vector with Stata display log
         if \code{nolog = FALSE}. May be an empty string if Stata does not finish executing
         the code before \code{timeout}.
         \item \code{error} -- If Stata displays an error -- an integer number with the Stata
         error code number (see \url{http://www.stata.com/manuals14/perror.pdf}).
         The error message should be visible in the \code{log} (see the point above).
         \item \code{df} -- Optional: a data frame saved by Stata if \code{import_df = TRUE} and
         if it can be read by \code{\link[utils]{read.delim}}, else \code{NULL}
         if \code{\link[utils]{read.delim}} returns an error.
			\item \code{results} -- if the argument \code{results} is not \code{NULL}, a list with
			one or both of the elements named \code{e_class} and/or \code{r_class}
			(depending on what was specified in the argument \code{results}), each including (if available):
			\itemize{
				   \item \code{scalars} -- a named list of numeric scalar values,
				   \item \code{macros} -- a named list of character (string/text) values,
				   \item \code{matrices} -- a named list of numeric matrices,
				   \item \code{modeldf} -- only for \code{e_class}: a data.frame with the estimated coefficients
				   (column \code{coef}), and standard errors (column \code{stderr}), with
				   the Stata variable names recorded in row.names.
			}
}
If \code{future = TRUE}, an object of S3 class 'StataFuture' to be used by
\code{\link[RStataLink]{getStataFuture}}.
}
\description{
Execute some Stata code in a running Stata instance
}

