% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{getStataFuture}
\alias{getStataFuture}
\title{Obtain Stata output from a job previously sent to Stata by \code{\link[RStataLink]{doInStata}}}
\usage{
getStataFuture(StataFuture)
}
\arguments{
\item{StataFuture}{An object of S3 class 'StataFuture' produced by
\code{\link[RStataLink]{doInStata}}, with argument \code{future = TRUE},
when a job was sent to Stata.}
}
\value{
See the description of the return value for \code{\link[RStataLink]{doInStata}} with
argument \code{future = FALSE}.
}
\description{
Obtain Stata output from a job previously sent to Stata by \code{\link[RStataLink]{doInStata}}
}

