% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{isStataReady}
\alias{isStataReady}
\title{Check if Stata instance is ready to receive a job (running and not busy)}
\usage{
isStataReady(id, timeout = 1)
}
\arguments{
\item{id}{An object of S3 class 'StataID' generated by \code{\link[RStataLink]{startStata}}}

\item{timeout}{Seconds to wait for Stata to respond. Default: 1.
Warning: too low value may result in ``false nagative'' i.e. non-busy Stata
instance not managing to reply on time; too high value may unnecessairly slow down
the execution of your R code -- setting up this parameter optimally may require
some trial and error based on your hardware
(setting \code{timeout} as low as possible without producing ``false positives'').}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if Stata instance is ready to receive a job (running and not busy)
}

