% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{startStata}
\alias{startStata}
\title{Open a Stata instance in a local server-like mode to be used by R through
\code{\link[RStataLink]{doInStata}}}
\usage{
startStata(timeout = 60, start_cmd = getOption("statapath"),
  compath = tempdir(), exit_on_error601 = FALSE,
  verify=TRUE)
}
\arguments{
\item{timeout}{Seconds to wait for the response/confirmation from Stata. Default: 60.}

\item{start_cmd}{A string specifying a system command line to start Stata, e.g.: on Windows OS
\code{'"C:\\\\Program Files\\\\Stata14\\\\StataMP.exe"'}.
Default: the contents of option "statapath" obtined with \code{getOption('statpath')}.}

\item{compath}{A path to a directory where the R<->Stata information exchange
sub-directory should be created. Default: \code{\link{tempdir}()}.}

\item{exit_on_error601}{Logical: should Stata be closed when the R--Stata data exchange
working directory disappears and Stata ``server'' stops
(\href{http://www.stata.com/search.cgi?query=601}{Stata error 601})? Default: \code{FALSE}.}

\item{verify}{Logical (default: \code{TRUE}). Should R wait for the confirmation that Stata
``server'' has been started. Use it as an escape hack if Stata starts, but the verification
cannot be obtained.}
}
\value{
An object of S3 class 'StataID' if Stata confirms that
the 'server' is ready, else an error.
}
\description{
A wrapper around Stata ``server'' code
}

