% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concurrent.R
\name{startStataCluster}
\alias{startStataCluster}
\title{Start a "cluster" of multiple Stata instances}
\usage{
startStataCluster(n = parallel::detectCores(), ...)
}
\arguments{
\item{n}{Number of instances. Default value determined by
\code{\link[parallel]{detectCores}}.}

\item{...}{Arguments passed to \code{\link[RStataLink]{startStata}}.}
}
\value{
A list of objects of S3 class 'StataID' if all Stata instances confirm that
they are ready, else an error.
}
\description{
A convenience wrapper around \code{\link[RStataLink]{startStata}}.
}

