% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{stopStata}
\alias{stopStata}
\title{Stop (close) Stata instance}
\usage{
stopStata(id, clear = FALSE)
}
\arguments{
\item{id}{An object of S3 class 'StataID' generated by \code{\link[RStataLink]{startStata}}.}

\item{clear}{Logical. Should the data in Stata memory be discarded
(i.e. Stata code \code{exit, STATA clear}). Default: \code{FALSE}.}
}
\value{
\code{NULL} or error if there is unsaved data in Stata memory and
\code{clear=FALSE}.
}
\description{
A convenience wrapper around Stata code \code{exit, STATA}.
}

