% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concurrent.R
\name{stopStataCluster}
\alias{stopStataCluster}
\title{Stop multiple Stata instances}
\usage{
stopStataCluster(cl, ...)
}
\arguments{
\item{cl}{A "cluster" -- a list of objects of S3 class 'StataID', possibly
generated by \code{\link[RStataLink]{startStataCluster}}.}

\item{...}{Other arguments passed to \code{\link[RStataLink]{stopStata}}.}
}
\description{
A convenience wrapper around \code{\link[RStataLink]{stopStata}}.
}

