% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{getStataFuture}
\alias{getStataFuture}
\title{Obtain Stata output from a job previously sent to Stata by \code{\link[RStataLink]{doInStata}}}
\usage{
getStataFuture(StataFuture)
}
\arguments{
\item{StataFuture}{An object of S3 class 'StataFuture' produced by
\code{\link[RStataLink]{doInStata}}, with argument \code{future = TRUE},
when a job was sent to Stata.}
}
\value{
A list with:
\itemize{
         \item \code{log} -- A character vector with Stata display log
         if \code{nolog} in the job request was is \code{FALSE}.
         May be an empty string if Stata does not finish executing
         the code before the \code{timeout} set in the job.
         \item \code{df} -- Optional: a data frame saved by Stata if
         \code{import_df} was \code{TRUE} in the job and
         if it can be read by \code{\link[utils]{read.delim}}, else \code{NULL}
         if \code{\link[utils]{read.delim}} returns an error.
}
}
\description{
Obtain Stata output from a job previously sent to Stata by \code{\link[RStataLink]{doInStata}}
}

